requirements.txt
streamlit
pandas
openai
python-dotenv
gspread
oauth2client
